// src/Components/pages/filtered-film-list-Components/types/enum.ts
export enum userSelector {
  DISABLE_USER_FILTER = 'DISABLE_USER_FILTER',
  FAVORITES = 'FAVORITES',
  WATCH_LATER = 'WATCH_LATER',
}

export enum ConditionsData {
  POPULARITY_DOWN = 'POPULARITY_DOWN',
  POPULARITY_UP = 'POPULARITY_UP',
  VOTE_AVERAGE_DOWN = 'VOTE_AVERAGE_DOWN',
  VOTE_AVERAGE_UP = 'VOTE_AVERAGE_UP',
}